refac: feature type file 통합
