+++ src/features/add-comment/model/index.ts
@@ -1,2 +1,1 @@
-export { useAddCommentForm } from "@/features/add-comment/model/use-add-comment-form"
-export type { AddCommentFormData, AddCommentCallbacks } from "@/features/add-comment/model/types"
+export { useAddCommentForm, type AddCommentFormData, type AddCommentCallbacks } from "./use-add-comment-form"
