--- src/features/add-comment/model/types.ts
@@ -1,10 +1,0 @@
-export interface AddCommentFormData {
-  body: string
-  postId: number
-  userId: number
-}
-
-export interface AddCommentCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
