+++ src/features/add-post/model/index.ts
@@ -1,2 +1,1 @@
-export { useAddPostForm } from "./use-add-post-form"
-export type { AddPostFormData, AddPostCallbacks } from "./types"
+export { useAddPostForm, type AddPostFormData, type AddPostCallbacks } from "./use-add-post-form"
