--- src/features/add-post/model/types.ts
@@ -1,10 +1,0 @@
-export interface AddPostFormData {
-  title: string
-  body: string
-  userId: number
-}
-
-export interface AddPostCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
