+++ src/features/edit-comment/model/index.ts
@@ -1,2 +1,1 @@
-export { useEditCommentForm } from "@/features/edit-comment/model/use-edit-comment-form"
-export type { EditCommentFormData, EditCommentCallbacks } from "@/features/edit-comment/model/types"
+export { useEditCommentForm, type EditCommentFormData, type EditCommentCallbacks } from "./use-edit-comment-form"
