--- src/features/edit-comment/model/types.ts
@@ -1,8 +1,0 @@
-export interface EditCommentFormData {
-  body: string
-}
-
-export interface EditCommentCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
