+++ src/features/edit-post/model/index.ts
@@ -1,2 +1,1 @@
-export { useEditPostForm } from "./use-edit-post-form"
-export type { EditPostFormData, EditPostCallbacks } from "./types"
+export { useEditPostForm, type EditPostFormData, type EditPostCallbacks } from "./use-edit-post-form"
