--- src/features/edit-post/model/types.ts
@@ -1,9 +1,0 @@
-export interface EditPostFormData {
-  title: string
-  body: string
-}
-
-export interface EditPostCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
