+++ src/features/like-comment/index.ts
@@ -1,2 +1,1 @@
 export { LikeCommentButton } from "./ui"
-export type { LikeCommentButtonProps, LikeCommentCallbacks } from "./ui"
