+++ src/features/like-comment/model/index.ts
@@ -1,2 +1,1 @@
-export { useLikeComment } from "./use-like-comment"
-export type { LikeCommentCallbacks } from "./types"
+export { useLikeComment, type LikeCommentCallbacks } from "./use-like-comment"
