--- src/features/like-comment/model/types.ts
@@ -1,4 +1,0 @@
-export interface LikeCommentCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
