+++ src/features/like-comment/ui/index.ts
@@ -1,2 +1,1 @@
 export { LikeCommentButton } from "./like-comment-button"
-export type { LikeCommentButtonProps } from "./like-comment-button"
