+++ src/features/like-comment/ui/like-comment-button.tsx
@@ -2,19 +2,14 @@ import React from "react"
 import { Button } from "@shared/ui"
 import { useLikeComment, type LikeCommentCallbacks } from "../model"
 
-export interface LikeCommentButtonProps {
+interface LikeCommentButtonProps {
   commentId: number
   postId: number
   currentLikes: number
   callbacks?: LikeCommentCallbacks
 }
 
-export const LikeCommentButton: React.FC<LikeCommentButtonProps> = ({
-  commentId,
-  postId,
-  currentLikes,
-  callbacks,
-}) => {
+export const LikeCommentButton: React.FC<LikeCommentButtonProps> = ({ commentId, postId, currentLikes, callbacks }) => {
   const { handleLike, isLoading } = useLikeComment()
 
   const onClick = () => {
@@ -22,12 +17,7 @@ export const LikeCommentButton: React.FC<LikeCommentButtonProps> = ({
   }
 
   return (
-    <Button 
-      variant="ghost" 
-      size="sm" 
-      onClick={onClick}
-      disabled={isLoading}
-    >
+    <Button variant="ghost" size="sm" onClick={onClick} disabled={isLoading}>
       좋아요 <span className="ml-1 text-xs">{currentLikes}</span>
     </Button>
   )
