+++ src/features/remove-comment/index.ts
@@ -1,2 +1,2 @@
 export { RemoveCommentButton } from "./ui"
-export type { RemoveCommentButtonProps, RemoveCommentCallbacks } from "./ui"
+export type { RemoveCommentButtonProps } from "./ui"
