--- src/features/remove-comment/model/types.ts
@@ -1,4 +1,0 @@
-export interface RemoveCommentCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
