--- src/features/remove-post/model/types.ts
@@ -1,4 +1,0 @@
-export interface DeletePostCallbacks {
-  onSuccess?: () => void
-  onError?: (error: unknown) => void
-}
